chore: add test that warn log contains webhook path
