+++ connector-runtime/spring-boot-starter-camunda-connectors/src/test/java/io/camunda/connector/runtime/inbound/WebhookControllerPlainJavaTests.java
@@ -24,13 +24,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.EvictingQueue;
-import io.camunda.connector.api.inbound.Activity;
-import io.camunda.connector.api.inbound.Severity;
 import io.camunda.connector.api.inbound.webhook.WebhookConnectorExecutable;
 import io.camunda.connector.api.inbound.webhook.WebhookProcessingPayload;
 import io.camunda.connector.api.inbound.webhook.WebhookResult;
@@ -43,14 +39,8 @@ import io.camunda.connector.runtime.inbound.lifecycle.ActiveInboundConnector;
 import io.camunda.connector.runtime.inbound.webhook.WebhookConnectorRegistry;
 import io.camunda.connector.validation.impl.DefaultValidationProvider;
 import java.util.Map;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
@@ -61,8 +51,6 @@ import org.mockito.quality.Strictness;
 public class WebhookControllerPlainJavaTests {
 
   private static final ObjectMapper mapper = ConnectorsObjectMapperSupplier.DEFAULT_MAPPER;
-
-  @Captor ArgumentCaptor<Activity> activityCaptor;
 
   @Test
   public void multipleWebhooksOnSameContextPathAreNotSupported() {
@@ -113,63 +101,6 @@ public class WebhookControllerPlainJavaTests {
 
     // then
     assertFalse(webhook.getWebhookConnectorByContextPath("myPath").isPresent());
-  }
-
-  private static Stream<Arguments> invalidCases() {
-    return Stream.of(
-        Arguments.of("&20encoded+whitespace"),
-        Arguments.of("€"),
-        Arguments.of("-my-path"),
-        Arguments.of("my_path_"));
-  }
-
-  private static Stream<Arguments> validCases() {
-    return Stream.of(
-        Arguments.of("z"),
-        Arguments.of("validAlphaOnly"),
-        Arguments.of("hello-world"),
-        Arguments.of("123-456_789"));
-  }
-
-  @ParameterizedTest
-  @MethodSource("invalidCases")
-  public void webhookPathLogIfInvalidCharacters(String webhookPath) {
-    WebhookConnectorRegistry webhook = new WebhookConnectorRegistry();
-
-    // given
-    var processA1 = buildConnector(webhookDefinition("processA", 1, webhookPath));
-
-    // when
-    webhook.register(processA1);
-
-    // then
-    verify(processA1.context(), times(1)).log(activityCaptor.capture());
-    assertEquals(Severity.WARNING, activityCaptor.getValue().severity());
-
-    assertWebhookRegistered(webhookPath, webhook, processA1);
-  }
-
-  private static void assertWebhookRegistered(
-      String webhookPath, WebhookConnectorRegistry webhook, ActiveInboundConnector processA1) {
-    var existingConnector = webhook.getWebhookConnectorByContextPath(webhookPath);
-    assertTrue(existingConnector.isPresent());
-    assertEquals(processA1, existingConnector.get());
-  }
-
-  @ParameterizedTest
-  @MethodSource("validCases")
-  public void webhookPathDontLogValidCharacters(String webhookPath) {
-    WebhookConnectorRegistry webhook = new WebhookConnectorRegistry();
-
-    // given
-    var processA1 = buildConnector(webhookDefinition("processA", 1, webhookPath));
-
-    // when
-    webhook.register(processA1);
-
-    // then
-    verify(processA1.context(), times(0)).log(activityCaptor.capture());
-    assertWebhookRegistered(webhookPath, webhook, processA1);
   }
 
   private static long nextProcessDefinitionKey = 0L;
